fix(cpu): now correctly cast immediate value as uint32_t before casting it to uint64_t in ADDIC instruction

+++ core/src/cpu/instructions/Add.cpp
@@ -89,7 +89,7 @@ namespace Core::Instruction {
     void ADDIC(Interpreter &cpu, const EncodedInstruction &instr)
     {
         const int32_t imm = static_cast<int16_t>(instr.si);
-        const uint64_t result = static_cast<uint64_t>(cpu.m_gpr[instr.ra]) + static_cast<uint64_t>(imm);
+        const uint64_t result = static_cast<uint64_t>(cpu.m_gpr[instr.ra]) + static_cast<uint64_t>(static_cast<uint32_t>(imm));
 
         cpu.m_gpr[instr.rt] = static_cast<uint32_t>(result);
         cpu.m_xer.ca = (result >> CARRY_OFFSET) & 1;
